JustFast
JustFast is a simple program based on FTP . It was useful for me to
transfer the android app made in eclipse  to mobile from ftp.
Any FTP program in android device that supports FTP can be helpful to
tranfer file. Something that we can see is the file transfer rate in
Byte, Kilo Byte and Mega Byte which may be very useful fro developing
process.